synth_rewrite complete
The synth module has been rewritten. Originally the synth generated sound by
emulating the hardware and collecting samples using nearest neighbor. While
this approach worked, the sound quality was poor at high frequencies due to
"naive" square waves being generated. The resulting waveforms were not
bandlimited, so there was ringing artifacts in the sound output, as well as
aliasing. In order to prevent this, bandlimited synthesis was used instead of
emulation. This technique is best explained here:
http://www.slack.net/~ant/bl-synth/

A lot of changes were made to the class hierarchy in the synth/ folder.
Specifically, the Channel base class and its subclasses were removed.
Individual hardware components were rewritten as separate classes, like
Envelope instead of EnvChannel and Sweep instead of SweepPulseChannel.
Channels can be made by using a combination of instances from of these
individual classes as opposed to inheriting from the Channel classes. This
makes the class design more flexible and easier to test/implement. For example,
channel 1 (formerly SweepPulseChannel) is now just a combination of instances
from the Envelope, Sweep and PulseOsc classes. The HardwareFile struct contains
the combinations for each channel.

The Mixer and Sequencer classes were merged into the Synth class, since these
were mainly used by the synth internally (except for setting the Mixer's
terminal states). Both components remain the same, except with a couple things
removed. The sequencer no longer sequences the length counters, as it was
never used by the tracker and most likely will not be needed. The mixer no
longer emulates terminal volume and panning, as the tracker will not need it
either (however this can be easily implemented by multiplying a gain to the
output buffer).

The bandlimited synthesis was implemented in the Osc base class. There are two
subclasses, PulseOsc and WaveOsc. The former generates square waveforms with 4
possible duty cycles. The latter can generate a waveform represented by 32
4-bit samples. Osc works by generating a period buffer, which can be sampled
from to fill a buffer of any size.

The synth generates float samples again, instead of 16-bit integers. Generated
waveforms now range from -1.0 to 1.0 instead of 0.0 to 1.0.