Renderer rewrite (callback is now lock-free)
Lots of changes here, but this commit mainly improves the Renderer
and gets it working for MDI. Before this commit the Renderer did all
its renderering in the callback thread. While this may seem like a
fine idea in the first place, doing it this way has a few
complications:
 - Since the callback is called from a real-time/high priority thread,
   all code in the callback must execute in a predictable/exact amount
   of time. So no blocking functions like I/O, memory allocation,
   syscalls, etc are allowed.
 - Mutexes cannot be used since they block/use syscalls, resulting in
   a creative approach to synchronization via spinlocks and atomic
   variables.
 - The callback cannot stop itself (doing so results in deadlock), so
   a separate thread running in the background was necessary to stop
   it when needed.

Now, the callback thread simply reads from a ringbuffer, which
is periodically replenished in a separate thread via the render
function. The process is simple:
 - The device callback reads samples from the ringbuffer when needed
 - The render function renders samples to this buffer, filling it
   completely. It is called periodically, via a QTimer (in its own
   thread).
 - This process repeats until there is nothing to render

Doing it this way makes it much easier and less complicated for the
renderering logic. It also allows us to use regular mutexes instead
of spinlocks. Thus eliminating the nightmare that is lockfree
synchronization. Downside is we have to be careful with timing.
Underruns will occur if the renderering thread cannot produce samples
faster than the callback thread consumes them. Prior to this,
underruns only occurred on failing to lock the spinlock.

Hopefully, this rewrite is effective and we won't have to revert
(again).